ease of use update
centralized movement, camera, and stamina settings to the root node to make it easier to use in a quick drop-in scenario.

Replaced "get_last_mouse_velocity" with "event.relative" since event.relative provides a more accurate and consistent output of mouse movements. get_last_mouse_velocity is not well suited for mose controlled cameras and is better for getting the calculated speed of the mouse and not necessarily it's actual change in position.